requirements.txt
black==19.10b0
colorlog==4.1.0
Django==3.0.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
pycrypto==2.6.1
mysqlclient
voluptuous==0.11.7
